# Set property directly, rather than using defineProperty