--- a/.travis.yml
+++ b/.travis.yml
@@ -5,7 +5,7 @@ script:
 - make test
 node_js:
 - '0.10'
-- '0.12'
+- '0.12.2'
 before_deploy:
 - npm-prepublish --verbose
 deploy: