--- a/lib/models/article.js
+++ b/lib/models/article.js
@@ -24,7 +24,16 @@ Article.prototype.parse = function (obj) {
 	this.id = obj.item.id;
 	this.raw = obj;
 
-	if (!this.raw.item.body) return;
+	if (this.raw.item.body) {
+		var html = this.raw.item.body.body;
+		try {
+			this.body = dom.removeNonArticleLinks(dom.fixRelativeLinks(dom.removeEmptyParagraphs(html)));
+		} catch (e) {
+			this.body = '<p>Error parsing this article.</p>';
+		}
+	} else {
+		this.body = '';
+	}
 };
 
 /**
@@ -466,25 +475,6 @@ Object.defineProperty(Article.prototype, 'quotes', {
 		} else {
 			return [];
 		}
-	}
-});
-
-/**
- * The article body as HTML.
- */
-Object.defineProperty(Article.prototype, 'body', {
-	get: function () {
-
-		var html = this.raw.item.body.body;
-
-		try {
-			return dom.removeNonArticleLinks(dom.fixRelativeLinks(dom.removeEmptyParagraphs(html)));
-		} catch (e) {
-			// console.error('Error parsing article body', e);
-			return '<p>Error parsing this article.</p>';
-		}
-
-
 	}
 });