--- a/lib/models/article.js
+++ b/lib/models/article.js
@@ -47,6 +47,9 @@ Article.prototype.paragraphs = function (to, from, options) {
 	return $('p').slice(to, from);
 };
 
+// TODO: move the following out of defineProperty and just set `this` directly
+// https://github.com/Financial-Times/ft-api-client/pull/136
+
 /**
  * Returns a list of inbound links to other FT articles contained within the article body
  */