Causing a big performance hit, e.g. every time `this.body` was being called, the function was run.

Starting with just this, but all the properties in here should set `this` directly

Using Engels as a test,

### Before

```
$ siege -c 12 -t 20s http://localhost:3002/uk
...
Transactions:		          40 hits
Availability:		      100.00 %
Elapsed time:		       19.40 secs
Data transferred:	       12.02 MB
Response time:		        4.67 secs
Transaction rate:	        2.06 trans/sec
Throughput:		        0.62 MB/sec
Concurrency:		        9.62
Successful transactions:          40
Failed transactions:	           0
Longest transaction:	        6.59
Shortest transaction:	        1.73
```

### After

```
$ siege -c 12 -t 20s http://localhost:3002/uk
...
Transactions:		          91 hits
Availability:		      100.00 %
Elapsed time:		       19.48 secs
Data transferred:	       27.36 MB
Response time:		        2.01 secs
Transaction rate:	        4.67 trans/sec
Throughput:		        1.40 MB/sec
Concurrency:		        9.39
Successful transactions:          91
Failed transactions:	           0
Longest transaction:	        2.80
Shortest transaction:	        1.21
```