@matthew-andrews 2015-06-02
wow, great work.  i never liked `defineProperty`.  too magic.