@wheresrhys 2015-06-03
I think defineProperty is still useful for expensive actions that don't need to be carried out for most instances, but the vast majority of the properties here don't meet this criteria