@ironsidevsquincy 2015-06-03
It's not `defineProperty` that's inherently slow, it's just every time a property is called it runs the function, only to return a constant value in our case. 

Particular case is `this.readingTime` was being called, which in turn was calling `this.wordCount`, which in turn was calling `this.body`, which was running the intensive `dom.removeNonArticleLinks(dom.fixRelativeLinks(dom.removeEmptyParagraphs(html)))` _every time_, even though these values don't change. And when you've got [this](https://github.com/Financial-Times/next-article-card-component/blob/master/templates/partials/readingTime.html), which was calling `readingTime` up to 3 times, you get a major slowdown

Basically, only use `defineProperty` if you _need_ to watch the setting/getting of a property, of the property changes